refactor: atom 상태정의 모듈 수정, storage utils 분리
