+++ src/pages/CreateTimeDivider/index.js
@@ -10,7 +10,7 @@ import Text from 'shared/components/Text'
 import TimeSelectForm from 'shared/components/TimeSelectForm'
 import { convertHourMinuteToSeconds, convertSecondsToHourMinute } from 'shared/utils/convertTime'
 import { useSetRecoilState } from 'recoil'
-import { timerObject, timerState } from 'atom'
+import { timerObject, timerState } from 'state/timer'
 import { themeColors } from 'shared/constants/colors'
 
 const BUTTON_TEXT = Object.freeze({
