+++ src/shared/components/AddFormModal.jsx
@@ -3,7 +3,7 @@ import { HOUR_NUMBERS, MINUTE_NUMBERS } from './TimeSelectForm'
 import FormModal from './FormModal'
 import React from 'react'
 import { useRecoilState } from 'recoil'
-import { addMode, defaultMode, modeState, timerObject, timerState } from 'atom'
+import { addMode, defaultMode, modeState, timerObject, timerState } from 'state/timer'
 import Input from './Input'
 import styled from 'styled-components'
 import Text from './Text'
