+++ src/shared/components/Timer.jsx
@@ -1,6 +1,6 @@
 import { useTimer } from 'react-timer-hook'
 import { useRecoilState } from 'recoil'
-import { timerState } from 'atom'
+import { timerState } from 'state/timer'
 import React, { useEffect } from 'react'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
