+++ src/state/user.js
@@ -1,0 +1,15 @@
+import { atom } from 'recoil'
+import { sessionStorageEffect } from './utils/storage'
+
+export const loginUserState = atom({
+	key: 'loginUser',
+	default: {
+		token: null,
+		user: {
+			email: '',
+			fullName: '',
+			posts: [],
+		},
+	},
+	effects: [sessionStorageEffect('loginUser')],
+})
